requirements.txt
alembic==0.9.6
appnope==0.1.0
APScheduler==3.6.3
arrow==0.15.5
asn1crypto==0.24.0
astroid==1.6.3
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bahttext==1.0.1
bleach==2.1.3
blinker==1.4
cachetools==2.1.0
cairocffi==0.9.0
CairoSVG==1.0.22
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
configparser==3.5.0
cryptography==2.2.2
cssselect2==0.2.2
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
Flask==1.0.2
Flask-Admin==1.5.1
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.8.0
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WeasyPrint==0.6
Flask-WTF==0.14.2
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.18.2
futures==3.2.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
gspread==3.0.0
gunicorn==19.7.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
infinity==1.4
intervals==0.8.1
ipaddress==1.0.22
ipykernel==4.8.2
ipython==5.8.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
jdcal==1.3
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyterlab==0.31.12
jupyterlab-launcher==0.10.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
line-bot-sdk==1.15.0
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
numpy==1.13.3
oauth2client==4.1.2
openpyxl==2.5.1
pandas==0.22.0
pandocfilters==1.4.2
pathlib2==2.3.0
pdfrw==0.4
pexpect==4.4.0
pickleshare==0.7.4
Pillow==6.2.1
ply==3.10
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDrive==1.3.1
Pygments==2.2.0
pylint==1.8.4
pyOpenSSL==17.5.0
pyparsing==2.2.0
PyPDF2==1.26.0
Pyphen==0.9.5
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.6.1
python-dotenv==0.10.3
python-editor==1.0.3
pytz==2017.3
PyYAML==5.3.1
pyzmq==17.0.0
reportlab==3.5.32
requests==2.18.4
rsa==3.4.2
scandir==1.7
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.3
subprocess32==3.2.7
terminado==0.8.1
testpath==0.3.1
tinycss2==0.6.1
tornado==5.0.1
traitlets==4.3.2
typing==3.7.4.1
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
validators==0.14.2
wcwidth==0.1.7
WeasyPrint==0.42
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
xlrd==1.1.0
XlsxWriter==1.2.8
